Pour une  utilisation optimal du model MVC je developpe les besoins client (fonctionnalites userCases) dans le package business (metier) *Refactoring the code with structure *adding new Spring Components Scanner int the appApplication my software entry point.

+++ consumer/src/main/java/dao/CommentaireDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentaireDao extends JpaRepository <Commentaire, Integer>{
+public interface CommentaireDao extends JpaRepository <Commentaire, Long>{
 
 }
